flask-backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
astor=0.8.0=py37_0
astroid=2.4.2=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.4=py_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
certifi=2020.4.5.2=py37_0
click=7.1.2=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
flask=1.1.2=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.3.3=py_0
gevent=20.6.2=pypi_0
glib=2.63.1=h3eb4bd4_1
gmp=6.1.2=h6c8ec71_1
greenlet=0.4.16=pypi_0
grpcio=1.27.2=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=py37h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=he6710b0_3
imageio=2.8.0=pypi_0
importlib-metadata=1.6.0=py37_0
importlib_metadata=1.6.0=0
intel-openmp=2020.1=217
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=pypi_0
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_0
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py37_0
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37hfd86e86_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.1=py37h0573a6f_0
mock=4.0.2=py_0
nbconvert=5.6.1=py37_0
nbformat=5.0.6=py_0
ncurses=6.2=he6710b0_1
networkx=2.4=pypi_0
notebook=6.0.3=py37_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.3=py37h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.1.2=py37hb39fc2d_0
pip=20.0.2=py37_3
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
protobuf=3.11.4=py37he6710b0_0
ptyprocess=0.6.0=py37_0
pygments=2.6.1=py_0
pylint=2.5.3=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.16.0=py37h7b6447c_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
pytz=2020.1=py_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
send2trash=1.5.0=py37_0
setuptools=46.4.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
sqlite=3.31.1=h62c20be_1
tensorboard=1.13.1=py37hf484d3e_0
tensorflow=1.13.1=gpu_py37hc158e3b_0
tensorflow-base=1.13.1=gpu_py37h8d69cac_0
tensorflow-estimator=1.13.0=py_0
tensorflow-gpu=1.13.1=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tifffile=2020.6.3=pypi_0
tk=8.6.8=hbc83047_0
toml=0.10.0=py37h28b3542_0
tornado=6.0.4=py37h7b6447c_1
traitlets=4.3.3=py37_0
typed-ast=1.4.1=py37h7b6447c_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.34.2=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zope-event=4.4=pypi_0
zope-interface=5.1.0=pypi_0
zstd=1.3.7=h0b5b093_0
